+++ packages/bevy/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
 name = "bevy_dioxus"
-version = "0.0.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+version = "0.0.1"
+edition = "2021"
 
 [dependencies]
 dioxus-core = { path = "../core", version = "^0.1.9" }
